feat: Universal architecture foundation (Phases 0-5.1)
Implement core infrastructure for universal desktop/mobile app architecture.

**Phase 0: Setup**
- Install dependencies: zustand, framer-motion, react-router-dom

**Phase 1: Project Restructure**
- Create new directory structure: core/, ui/, screens/
- Move files to new structure (preserving git history):
  - types.ts → core/types/
  - api.ts → core/animepahe/
  - utils → core/utils/
  - hooks → ui/hooks/
  - components/ui → ui/components/base/
- Create centralized store exports

**Phase 2: State Management**
- Implement Zustand stores with persist middleware:
  - PreferenceStore: Settings management with Tauri persistence
  - QueueStore: Download queue with event listeners
  - LibraryStore: Library entries management
  - NetworkStore: Network status and mobile data detection
- Event-driven architecture for download progress updates

**Phase 3: Responsive Design System**
- Update Tailwind config with custom breakpoints
- Add boxShadow.glow and spacing.safe utilities
- Create useBreakpoint hook with SSR guards
- Build mobile-first components:
  - BottomNav: 3-tab mobile navigation
  - Sheet: Mobile-friendly modal component
  - ContextFAB: Floating action button

**Phase 4: Routing Infrastructure**
- Setup React Router v6 with BrowserRouter
- Create AppChrome layout wrapper
- Build DesktopHeader navigation component
- Implement responsive layout orchestration
- Configure routes for Download, Library, Settings

**Phase 5.1: Download Screen (Initial)**
- Extract Download screen with proper structure
- Establish component architecture
- Add comprehensive TODO comments for full migration
- Preserve App.old.tsx for reference

**Technical Highlights:**
- Responsive breakpoint at md (768px)
- Zustand persist with Tauri storage adapter
- SSR-safe hooks for Tauri environment
- Git history preserved for all moved files
- Build tested and passing
