+++ package.json
@@ -40,17 +40,21 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "crypto-js": "^4.2.0",
+    "framer-motion": "^12.23.24",
     "lucide-react": "^0.358.0",
     "posthog-js": "^1.268.9",
     "react": "^18.3.0",
     "react-dom": "^18.3.0",
-    "tailwind-merge": "^2.2.1"
+    "react-router-dom": "^7.9.4",
+    "tailwind-merge": "^2.2.1",
+    "zustand": "^5.0.8"
   },
   "devDependencies": {
     "@tauri-apps/cli": "^2",
     "@types/node": "^20.12.7",
     "@types/react": "^18.2.39",
     "@types/react-dom": "^18.2.17",
+    "@types/react-router-dom": "^5.3.3",
     "@vitejs/plugin-react": "^4.2.0",
     "autoprefixer": "^10.4.17",
     "ffmpeg-static": "^5.2.0",
