+++ src/core/store/index.ts
@@ -1,0 +1,16 @@
+/**
+ * Central store exports for state management
+ * Uses Zustand for reactive state across the application
+ */
+
+export { usePreferenceStore } from './preference-store';
+export { useQueueStore } from './queue-store';
+export { useLibraryStore } from './library-store';
+export { useNetworkStore } from './network-store';
+
+export type {
+  PreferenceState,
+  QueueState,
+  LibraryState,
+  NetworkState,
+} from './types';
