+++ src/core/utils/updateChecker.ts
