+++ src/screens/library/LibraryScreen.tsx
@@ -1,0 +1,16 @@
+/**
+ * Library Screen
+ * Shows downloaded anime and episodes
+ * Will be migrated from LibraryView.tsx in Phase 5.2
+ */
+export function LibraryScreen() {
+  return (
+    <div className="p-6">
+      <h1 className="text-2xl font-bold">Library Screen</h1>
+      <p className="text-muted-foreground mt-2">
+        This screen will display your library of downloaded anime.
+        Coming in Phase 5.2...
+      </p>
+    </div>
+  );
+}
