+++ src/screens/settings/SettingsScreen.tsx
@@ -1,0 +1,16 @@
+/**
+ * Settings Screen
+ * Application settings and preferences
+ * Will be created in Phase 5.3
+ */
+export function SettingsScreen() {
+  return (
+    <div className="p-6">
+      <h1 className="text-2xl font-bold">Settings Screen</h1>
+      <p className="text-muted-foreground mt-2">
+        This screen will contain app settings, theme preferences, and configuration.
+        Coming in Phase 5.3...
+      </p>
+    </div>
+  );
+}
