+++ src/ui/components/base/card.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "../../core/utils/cn";
+import { cn } from "../../../core/utils/cn";
 
 function Card({ className, ...props }: React.HTMLAttributes<HTMLDivElement>) {
   return (
