+++ src/ui/components/base/dialog.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as DialogPrimitive from "@radix-ui/react-dialog";
 
-import { cn } from "../../core/utils/cn";
+import { cn } from "../../../core/utils/cn";
 
 const Dialog = DialogPrimitive.Root;
