+++ src/ui/components/base/dropdown-menu.tsx
@@ -2,7 +2,7 @@ import * as React from "react"
 import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu"
 import { Check, ChevronRight, Circle } from "lucide-react"
 
-import { cn } from "../../core/utils/cn"
+import { cn } from "../../../core/utils/cn"
 
 const DropdownMenu = DropdownMenuPrimitive.Root
