+++ src/ui/components/base/input.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "../../core/utils/cn";
+import { cn } from "../../../core/utils/cn";
 
 export interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {}
