+++ src/ui/components/base/select.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import * as SelectPrimitive from "@radix-ui/react-select";
 import { Check, ChevronDown, ChevronUp } from "lucide-react";
 
-import { cn } from "../../core/utils/cn";
+import { cn } from "../../../core/utils/cn";
 
 const Select = SelectPrimitive.Root;
