+++ src/ui/components/base/tabs.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import * as TabsPrimitive from "@radix-ui/react-tabs"
 
-import { cn } from "../../core/utils/cn"
+import { cn } from "../../../core/utils/cn"
 
 const Tabs = TabsPrimitive.Root
