+++ src/ui/hooks/useAutoResumeDetection.ts
